backend/requirements.txt
alpaca-py==0.43.2
chromadb==1.5.6
duckduckgo_search==8.1.1
fastapi==0.135.3
google-genai==1.70.0
langchain-chroma==1.1.0
langchain-community==0.4.1
langchain-core==1.2.27
langchain-google-genai==4.2.1
langgraph==1.1.6
pandas==3.0.2
pandas-ta==0.4.71b0
python-dotenv==1.2.2
tenacity==9.1.4
transformers==5.5.0
uvicorn==0.44.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
pydantic==2.12.5
pydantic-settings==2.13.1
httpx==0.28.1
aiohttp==3.13.5
torch==2.11.0
numpy==2.2.6
numba==0.61.2
